feat: kick off refactor of optimization flow
