+++ pypackages/risk_engine/src/optimization/runRiskOptimization.py
@@ -1,13 +1,11 @@
-import json
 import os
 import numpy as np
-import optuna
 import pandas as pd
 from risk_engine.src.calculators.RiskMetrics import RiskMetrics as rm
 from risk_engine.tests.mocks.mockPosition import mock_position
 from risk_engine.src.constants import YEAR_IN_SECONDS
 from risk_engine.src.simulations.margin_requirements.MarginRequirements import MarginRequirements
-
+from risk_engine.src.optimization.runParameterOptimization import run_param_optimization
 
 # todo: turn these into an argument for generate pool function -> stateless
 MARKET = "irs_usdc"
@@ -15,76 +13,6 @@ RUN_OPTUNA = True
 RUN_SIMPLER_OPTIMISATION = True
 positions = mock_position[MARKET]
 
-
-def generate_pool(
-    df, name, p_lm, gamma, lambda_taker, lambda_maker, spread, lookback, min_leverage=20
-) -> float:  # Populate with risk parameters
-    # df: need to pass the oracle to the end-to-end test in the correct way
-    print(df)
-    mean_apy = df["apy"].mean()
-    if spread >= mean_apy:
-        spread -= 0.001
-    std_dev = df["apy"].std()
-    duration = df["timestamp"].values[-1] - df["timestamp"].values[0]
-    # Build the input of the simulation
-    risk_parameter = std_dev * np.sqrt(YEAR_IN_SECONDS / duration) * p_lm
-
-    # Instantiate the IRS pool and simulation
-    simulation = MarginRequirements()
-
-    print("TIMESTAMPS: ", df["timestamp"].values)
-    print("INDEX: ", df["liquidity_index"].values)
-
-    # TODO: how can the fixed rate lookback window be added here i.e. where is the GWAP treated?
-    # I think this needs to be added to the IRS market object
-    simulation.setUp(
-        collateral_token=positions["base_token"],
-        initial_fixed_rate=mean_apy,
-        risk_parameter=risk_parameter,
-        im_multiplier=gamma,
-        slippage_phi=positions["phi"],
-        slippage_beta=positions["beta"],
-        lp_spread=spread,
-        is_trader_vt=True,
-        timestamps=df["timestamp"].values,
-        indices=df["liquidity_index"].values,
-        maker_fee=lambda_maker,
-        taker_fee=lambda_taker,
-        gwap_lookback=lookback,
-    )
-
-    """
-    1. Agent-based simulation of maker and taker positions in the
-       IRS pool
-    """
-    simulation_folder = f"./{MARKET}/{name}/optuna/"
-    if not os.path.exists(simulation_folder):
-        os.makedirs(simulation_folder)
-    output = simulation.run(output_folder=simulation_folder)  # Add a return output to sim.run
-
-    """
-    2. Optimisation function implementation for Optuna
-    """
-    # We need to evaluate the objective function here
-    # We use the lists defined above here
-    average_leverage = 0.5 * (
-        positions["maker_amount"] / output.iloc[0]["lp_liquidation_threshold"]
-        + positions["taker_amount"] / output.iloc[0]["trader_liquidation_threshold"]
-    )
-    average_risk = 0.5 * (
-        output["lp_uPnL"].std() + output["trader_uPnL"].std()
-    )  # Normalise risk to collateral supplied
-    regularisation = (
-        10 if average_leverage < min_leverage else 0
-    )  # Add in a regularisation term to constrain the minimum leverage
-
-    # Finally we need to compute an insolvency VaR, IVaR
-    risk_metrics = rm.RiskMetrics(df=output)
-    lvar, ivar = risk_metrics.lvar_and_ivar()
-    ivar_reg = 10 if ivar < 0.95 else 0
-
-    objective = average_leverage - average_risk - regularisation - ivar_reg
-    return objective  # Objective function
 
 
 def main(p_lm, gamma, lambda_taker, lambda_maker, spread, lookback):
@@ -155,45 +83,8 @@ def objective(trial):
     return obj
 
 
-def run_param_optimization(parser):
-
-    parser.add_argument(
-        "-n_trials", "--n_trials", type=float, help="Number of optimization trials", default=2
-    )
-    parser.add_argument("-d", "--debug", action="store_true", help="Debug mode", default=False)
-    n_trials = parser.parse_args().n_trials
-
-    study = optuna.create_study(
-        direction="maximize",
-        sampler=optuna.samplers.TPESampler(),
-        pruner=optuna.pruners.SuccessiveHalvingPruner(),
-    )
-    study.optimize(objective, n_trials=n_trials)
-
-    # Relevant output plots
-    out_dir = f"./{MARKET}/optuna_final/"
-    if not os.path.exists(out_dir):
-        os.makedirs(out_dir)
-
-    # Output optimised results
-    trial = study.best_trial
-    print(f"Best optimised value: {trial.value}")
-
-    print("Optimised parameters: ")
-    for key, value in trial.params.items():
-        print(f"{key}: {value}")
-
-    fig = optuna.visualization.plot_optimization_history(study)
-    fig.write_image(out_dir + f"optuna_history_{MARKET}.png")
-
-    with open(out_dir + f"optimised_parameters_{MARKET}.json", "w") as fp:
-        json.dump(trial.params, fp, indent=4)
-
-
 if __name__ == "__main__":
-    # Adding an argument parser
     from argparse import ArgumentParser
-
     parser = ArgumentParser()
 
     if RUN_OPTUNA:
