+++ pypackages/risk_engine/tests/end_to_end/optimization/run_risk_parametrization_test.py
@@ -1,0 +1,9 @@
+import unittest
+from risk_engine.src.optimization.runRiskOptimization import run_param_optimization
+
+class RunRiskOptimizationTest(unittest.TestCase):
+
+    def setUp(self):
+
+        pass
+
